# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Drifter Highway Space API",
        "description": "EVE Online authentication and character management API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local development server"
        },
        {
            "url": "https://drifterhighway.space",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/characters": {
            "get": {
                "tags": ["Characters"],
                "summary": "Get All Characters",
                "description": "Retrieve all active characters from the database.",
                "operationId": "GetCharacters",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Which page of results to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "How many results to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of characters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Character"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - valid JWT token required"
                    },
                    "404": {
                        "description": "No characters found"
                    }
                }
            },
            "put": {
                "tags": ["Characters"],
                "summary": "Update Character",
                "description": "Update a character by ID.",
                "operationId": "UpdateCharacter",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "description": "Character ID"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CharacterUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Character updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Character"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Characters"],
                "summary": "Delete Character",
                "description": "Delete (deactivate) a character by ID. Character is marked as Inactive.",
                "operationId": "DeleteCharacter",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "description": "Character ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Character deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/characters/{id}": {
            "get": {
                "tags": ["Characters"],
                "summary": "Get Character by ID",
                "description": "Retrieve a specific character by ID.",
                "operationId": "GetCharacter",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "description": "Character ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Character found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Character"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "tags": ["Characters"],
                "summary": "Update Character by ID",
                "description": "Update a specific character.",
                "operationId": "UpdateCharacterById",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "description": "Character ID"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CharacterUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Character updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Character"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Characters"],
                "summary": "Delete Character by ID",
                "description": "Delete (deactivate) a specific character.",
                "operationId": "DeleteCharacterById",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "description": "Character ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Character deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/corporations": {
            "get": {
                "tags": ["Corporations"],
                "summary": "Get All Corporations",
                "description": "Retrieve all corporations from the database.",
                "operationId": "GetCorporations",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Which page of results to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "How many results to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of corporations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Corporation"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - valid JWT token required"
                    },
                    "404": {
                        "description": "No corporations found"
                    }
                }
            },
            "put": {
                "tags": ["Corporations"],
                "summary": "Update Corporation",
                "description": "Update a corporation by ID.",
                "operationId": "UpdateCorporation",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "description": "Corporation ID"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CorporationUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Corporation updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Corporation"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Corporations"],
                "summary": "Delete Corporation",
                "description": "Delete (deactivate) a corporation by ID.",
                "operationId": "DeleteCorporation",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "description": "Corporation ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Corporation deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/corporations/{id}": {
            "get": {
                "tags": ["Corporations"],
                "summary": "Get Corporation by ID",
                "description": "Retrieve a specific corporation by ID.",
                "operationId": "GetCorporation",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "description": "Corporation ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Corporation found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Corporation"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "tags": ["Corporations"],
                "summary": "Update Corporation by ID",
                "description": "Update a specific corporation.",
                "operationId": "UpdateCorporationById",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "description": "Corporation ID"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CorporationUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Corporation updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Corporation"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Corporations"],
                "summary": "Delete Corporation by ID",
                "description": "Delete (deactivate) a specific corporation.",
                "operationId": "DeleteCorporationById",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "description": "Corporation ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Corporation deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Bearer Token for authentication"
            }
        },
        "schemas": {
...
            "Character": {
                "type": "object",
                "description": "Character entity from the database",
                "properties": {
                    "ID": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Unique identifier for the character"
                    },
                    "Name": {
                        "type": "string",
                        "description": "Name of the character"
                    },
                    "User": {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID of the user who owns this character"
                    },
                    "Corporation": {
                        "type": "object",
                        "description": "Associated corporation",
                        "properties": {
                            "ID": {
                                "type": "integer",
                                "format": "int64",
                                "description": "Corporation ID"
                            },
                            "Name": {
                                "type": "string",
                                "description": "Corporation name"
                            }
                        }
                    },
                    "Status": {
                        "type": "string",
                        "description": "Status of the character",
                        "enum": [
                            "Active",
                            "Inactive",
                            "InvalidToken",
                            "Pending"
                        ]
                    },
                    "RefreshToken": {
                        "type": "string",
                        "description": "Optional refresh token for authentication"
                    },
                    "AccessToken": {
                        "type": "string",
                        "description": "Optional access token for authentication"
                    },
                    "Ship": {
                        "type": "object",
                        "description": "Optional ship owned by this character",
                        "properties": {
                            "ID": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "Name": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "CharacterUpdate": {
                "type": "object",
                "description": "Request body for updating a character",
                "properties": {
                    "Name": {
                        "type": "string",
                        "description": "New name for the character"
                    }
                }
            },
            "Corporation": {
                "type": "object",
                "description": "Corporation entity from the database",
                "properties": {
                    "ID": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Unique identifier for the corporation"
                    },
                    "Name": {
                        "type": "string",
                        "description": "Name of the corporation"
                    },
                    "Type": {
                        "type": "string",
                        "description": "Type of group entity (Corporation/Alliance)",
                        "enum": ["Corporation", "Alliance"]
                    },
                    "Parent": {
                        "type": "object",
                        "description": "Optional parent group entity reference",
                        "properties": {
                            "ID": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "Name": {
                                "type": "string"
                            },
                            "Type": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "CorporationUpdate": {
                "type": "object",
                "description": "Request body for updating a corporation",
                "properties": {
                    "Name": {
                        "type": "string",
                        "description": "New name for the corporation"
                    }
                }
            }
        }
    }
}